cohortrag_engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cohortrag-engine"
version = "1.0.0"
description = "The Production-Ready Open Source RAG for Education"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "CohortRAG Team", email = "maintainers@cohortrag.com"}
]
maintainers = [
    {name = "CohortRAG Team", email = "maintainers@cohortrag.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Environment :: Web Environment",
]
keywords = [
    "rag", "retrieval-augmented-generation", "education", "ai", "llm",
    "vector-search", "educational-technology", "machine-learning",
    "natural-language-processing", "embeddings", "chatbot", "question-answering"
]
requires-python = ">=3.8"

dependencies = [
    # Core RAG Framework
    "llama-index>=0.10.0,<0.12.0",
    "llama-index-llms-gemini>=0.1.0",
    "llama-index-embeddings-huggingface>=0.1.0",

    # Vector Database
    "llama-index-vector-stores-chroma>=0.1.0",
    "chromadb>=0.4.0,<0.6.0",

    # Document Processing
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "python-dotenv>=1.0.0",

    # Advanced Features
    "sentence-transformers>=2.6.0",
    "FlagEmbedding>=1.2.0",

    # Evaluation Framework
    "ragas>=0.1.0",

    # Core Utilities
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",

    # Async and Performance
    "asyncio-throttle>=1.0.0",
    "psutil>=5.9.0",

    # Caching
    "redis>=5.0.0",

    # Logging and Monitoring
    "structlog>=23.0.0",

    # Web Framework (for future API)
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

production = [
    "prometheus-client>=0.17.0",
    "gunicorn>=21.0.0",
    "redis[hiredis]>=5.0.0",
]

docker = [
    "prometheus-client>=0.17.0",
    "redis[hiredis]>=5.0.0",
]

all = [
    "cohortrag-engine[dev,test,production]"
]

[project.urls]
Homepage = "https://github.com/YourUsername/CohortHelperAI"
Documentation = "https://github.com/YourUsername/CohortHelperAI/tree/main/cohortrag_engine/docs"
Repository = "https://github.com/YourUsername/CohortHelperAI.git"
"Bug Reports" = "https://github.com/YourUsername/CohortHelperAI/issues"
"Feature Requests" = "https://github.com/YourUsername/CohortHelperAI/issues"
Changelog = "https://github.com/YourUsername/CohortHelperAI/releases"
Discussions = "https://github.com/YourUsername/CohortHelperAI/discussions"

[project.scripts]
cohortrag = "cohortrag_engine.main:main"
cohortrag-server = "cohortrag_engine.server:run_server"
cohortrag-benchmark = "cohortrag_engine.cli.benchmark:main"
cohortrag-validate = "cohortrag_engine.cli.validate:main"

[project.entry-points."cohortrag_engine.retrievers"]
chroma = "cohortrag_engine.core.retrieval:CohortRAGRetriever"
cached = "cohortrag_engine.core.cached_retrieval:ProductionCohortRAGRetriever"

[project.entry-points."cohortrag_engine.vector_stores"]
chroma = "cohortrag_engine.core.retrieval:CohortRAGRetriever"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["cohortrag_engine*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
cohortrag_engine = [
    "*.json",
    "*.yaml",
    "*.yml",
    "templates/*.html",
    "static/*",
    "py.typed",
]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["cohortrag_engine"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# MyPy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual typing
ignore_missing_imports = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "llama_index.*",
    "chromadb.*",
    "ragas.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=cohortrag_engine",
    "--cov-branch",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
]
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests that require GPU",
    "expensive: marks tests that are computationally expensive",
]

# Coverage configuration
[tool.coverage.run]
source = ["cohortrag_engine"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/venv/*",
    "*/.venv/*",
    "*/build/*",
    "*/dist/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Ruff configuration (alternative to flake8)
[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]